ToggleFit implementation
When each image is switched to, the initial state
of whether to fit to screen or use action size
of image is computed.

Actual size is used if both dimensions of the image are
smaller than the canvas viewport. Otherwise, the image
is fit to screen.

The state can be toggled by pressing `z` for each image.
This change is reset upon changing images.